Rename `controller` to `change_notifier`
