+++ integration_test/wallet_test.dart
@@ -8,8 +8,8 @@ import 'package:ten_ten_one/balance.dart';
 import 'package:ten_ten_one/wallet/wallet_dashboard.dart';
 import 'package:ten_ten_one/wallet/wallet_lightning.dart';
 import 'package:ten_ten_one/model/amount.dart';
-import 'package:ten_ten_one/controller/balance_change_notifier.dart';
-import 'package:ten_ten_one/controller/seed_backup_change_notifier.dart';
+import 'package:ten_ten_one/change_notifier/balance_change_notifier.dart';
+import 'package:ten_ten_one/change_notifier/seed_backup_change_notifier.dart';
 import 'package:ten_ten_one/wallet/seed.dart';
 
 final GoRouter _router = GoRouter(
