+++ lib/cfd_trading/cfd_order_confirmation.dart
@@ -6,7 +6,7 @@ import 'package:intl/intl.dart';
 import 'package:provider/provider.dart';
 import 'package:ten_ten_one/cfd_trading/cfd_trading.dart';
 import 'package:ten_ten_one/model/amount.dart';
-import 'package:ten_ten_one/controller/cfd_trading_change_notifier.dart';
+import 'package:ten_ten_one/change_notifier/cfd_trading_change_notifier.dart';
 import 'package:ten_ten_one/model/order.dart';
 import 'package:ten_ten_one/utilities/tto_table.dart';
