+++ lib/change_notifier/cfd_trading_change_notifier.dart
