+++ lib/main.dart
@@ -11,16 +11,16 @@ import 'package:ten_ten_one/cfd_trading/cfd_order_confirmation.dart';
 import 'package:ten_ten_one/cfd_trading/cfd_order_detail.dart';
 import 'package:ten_ten_one/cfd_trading/cfd_trading.dart';
 import 'package:ten_ten_one/model/payment.dart';
-import 'package:ten_ten_one/controller/payment_history_change_notifier.dart';
+import 'package:ten_ten_one/change_notifier/payment_history_change_notifier.dart';
 import 'package:ten_ten_one/wallet/deposit.dart';
 import 'package:ten_ten_one/wallet/open_channel.dart';
 import 'package:ten_ten_one/wallet/wallet.dart';
-import 'package:ten_ten_one/controller/wallet_change_notifier.dart';
+import 'package:ten_ten_one/change_notifier/wallet_change_notifier.dart';
 import 'package:ten_ten_one/model/amount.dart';
-import 'package:ten_ten_one/controller/balance_change_notifier.dart';
-import 'package:ten_ten_one/controller/cfd_trading_change_notifier.dart';
+import 'package:ten_ten_one/change_notifier/balance_change_notifier.dart';
+import 'package:ten_ten_one/change_notifier/cfd_trading_change_notifier.dart';
 import 'package:ten_ten_one/model/order.dart';
-import 'package:ten_ten_one/controller/seed_backup_change_notifier.dart';
+import 'package:ten_ten_one/change_notifier/seed_backup_change_notifier.dart';
 import 'package:ten_ten_one/wallet/receive.dart';
 import 'package:ten_ten_one/wallet/seed.dart';
 import 'package:go_router/go_router.dart';
