+++ lib/wallet/wallet_dashboard.dart
@@ -3,10 +3,10 @@ import 'package:provider/provider.dart';
 import 'package:go_router/go_router.dart';
 import 'package:ten_ten_one/balance.dart';
 import 'package:ten_ten_one/cfd_trading/cfd_trading.dart';
-import 'package:ten_ten_one/controller/balance_change_notifier.dart';
-import 'package:ten_ten_one/controller/seed_backup_change_notifier.dart';
+import 'package:ten_ten_one/change_notifier/balance_change_notifier.dart';
+import 'package:ten_ten_one/change_notifier/seed_backup_change_notifier.dart';
 import 'package:ten_ten_one/model/service.dart';
-import 'package:ten_ten_one/controller/payment_history_change_notifier.dart';
+import 'package:ten_ten_one/change_notifier/payment_history_change_notifier.dart';
 import 'package:ten_ten_one/wallet/deposit.dart';
 import 'package:ten_ten_one/wallet/payment_history_list_item.dart';
 import 'package:ten_ten_one/wallet/seed.dart';
